% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dis.R
\name{convert_dis}
\alias{convert_dis}
\title{Takes a vector of CDE disability codes and converts them to the abbreviated disability name}
\usage{
convert_dis(x)
}
\arguments{
\item{x}{A numeric or character vector of disability codes}
}
\value{
A character vector of disability names
}
\description{
Takes a vector of CDE disability codes and converts them to the abbreviated disability name
}
