% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_subgroup.R
\name{convert_subgroup}
\alias{convert_subgroup}
\title{Takes a vector of CA Dashboard subgroup codes
and converts them to the subgroup name}
\usage{
convert_subgroup(x)
}
\arguments{
\item{x}{A character vector of CA Dashboard subgroup codes}
}
\value{
A character vector of subgroup names
}
\description{
Takes a vector of CA Dashboard subgroup codes
and converts them to the subgroup name
}
