% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sizescore.R
\name{make_sizescore}
\alias{make_sizescore}
\title{Plots subgroup size against distance from three score}
\usage{
make_sizescore(ela, math)
}
\arguments{
\item{ela}{Path for CA School Dashboard ELA dataset}

\item{math}{Path for CA School Dashboard math dataset}
}
\value{
Plot of subgroup size vs distance from three score
}
\description{
Plots subgroup size against distance from three score
}
