% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_a.R
\name{make_table_a}
\alias{make_table_a}
\title{Makes a CASEMIS Table A with randomized made-up names, IDs, schools, and districts}
\usage{
make_table_a(rows)
}
\arguments{
\item{rows}{A number that is equal to the desired number of students}
}
\value{
A tibble that is a practice CASEMIS Table A
}
\description{
Makes a CASEMIS Table A with randomized made-up names, IDs, schools, and districts
}
